Use pinned memory to store Box device bounds
The managed memory approach resulted in a significant number of page faults due
to the constant number of box updates. So the new strategy is to use pinned memory
to hold the device version of the Box lo and hi bounds. This means that all of the
CPU updates do not cause a page fault, but it also means that the device cannot
by default see the data. This is resolved using the cudaHostGetDevicePointer API call.
However, we cannot in general assume that the user wants the device version when they
call loVect() and hiVect(); they may do some operations prior to a device kernel, and
we also need the CPU version of this data to calculate the thread and block layout
for the CUDA kernel launch. So new functions, loVectF() and hiVectF(), are added to both
Box and BaseFab. These return pointers to the data that are specifically intended for
Fortran evaluations, and for the CUDA version return the device pointer to the (pinned)
host data. On the CPU they return the normal loVect() and hiVect() result.

The most important refactor needed because of this is that the data needs to be present
on the device for the lifetime of the kernel launch. So boxes created to store the lo
and hi bounds of a loop need to be created outside of a MFIter loop.